requirements.txt
numpy==1.26.4
opencv-python==4.8.0.76
packaging==25.0
pillow==12.0.0
PyMuPDF==1.26.5
easyocr==1.7.1
pyzbar==0.1.9
qrcode==8.2
rembg
onnxruntime
